+++ genomics_data_index/api/query/impl/TreeSamplesQuery.py
@@ -4,7 +4,7 @@ import abc
 import logging
 from typing import Union, List, cast
 
-from ete3 import Tree, TreeStyle
+from ete3 import Tree, TreeStyle, NodeStyle
 
 from genomics_data_index.api.query.SamplesQuery import SamplesQuery
 from genomics_data_index.api.query.impl.WrappedSamplesQuery import WrappedSamplesQuery
@@ -111,6 +111,49 @@ class TreeSamplesQuery(WrappedSamplesQuery, abc.ABC):
         tree.set_outgroup(sample_name)
         return self._create_from_tree_internal(tree)
 
+    def prune(self, preserve_branch_length: bool = True, include_present: bool = True,
+              include_unknown: bool = False, include_absent: bool = False) -> SamplesQuery:
+        """
+        Prunes the tree joined to this TreeSamplesQuery down to only the currently selected samples.
+        :param preserve_branch_length: True if branch lengths between the subset of samples should be preserved, False otherwise.
+        :param include_present: If samples with a present status should be included in the tree.
+        :param include_unknown: If samples with an unknown status should be included in the tree.
+        :param include_absent: If samples with an absent status from query should be included in the tree.
+        :return: A new TreeSamplesQuery with the tree pruned down to only the selected samples (and the reference
+                 genome if it is one of the leaves of the tree).
+        """
+        tree = self._tree_copy_prune(preserve_branch_length=preserve_branch_length,
+                                     include_present=include_present, include_unknown=include_unknown,
+                                     include_absent=include_absent)
+        return self._create_from_tree_internal(tree)
+
+    def _tree_copy_prune(self,
+                         preserve_branch_length: bool,
+                         include_present: bool,
+                         include_unknown: bool,
+                         include_absent: bool,
+                         from_query: SamplesQuery = None,
+                         ) -> Tree:
+        tree = self._tree_copy()
+        if from_query is not None:
+            query = from_query
+        else:
+            query = self
+
+        nodes_to_keep = self._get_samples_to_keep_from_query(query, include_present=include_present,
+                                                             include_unknown=include_unknown,
+                                                             include_absent=include_absent)
+        tree.prune(nodes_to_keep, preserve_branch_length=preserve_branch_length)
+        return tree
+
+    def _get_samples_to_keep_from_query(self, query: SamplesQuery,
+                                        include_present: bool,
+                                        include_unknown: bool,
+                                        include_absent: bool,
+                                        ) -> List[str]:
+        return query.tolist(names=True, include_present=include_present, include_unknown=include_unknown,
+                            include_absent=include_absent)
+
     @abc.abstractmethod
     def _create_from_tree_internal(self, tree: Tree) -> SamplesQuery:
         pass
@@ -145,6 +188,7 @@ class TreeSamplesQuery(WrappedSamplesQuery, abc.ABC):
                     initial_style: TreeStyle = None,
                     mode='r',
                     highlight_style: Union[str, HighlightStyle] = 'pastel',
+                    node_style: NodeStyle = None,
                     legend_nsize: int = 20, legend_fsize: int = 11,
                     annotate_color_present: str = 'black',
                     annotate_color_absent: str = 'white',
@@ -172,8 +216,12 @@ class TreeSamplesQuery(WrappedSamplesQuery, abc.ABC):
                     include_unknown: bool = True,
                     show_leaf_names: bool = True,
                     show_legend_type_labels: bool = True,
-                    legend_type_label_present: str = 'P',
-                    legend_type_label_unknown: str = 'U',
+                    legend_type_label_present: str = 'Pr.',
+                    legend_type_label_unknown: str = 'Un.',
+                    rotation: float = 0,
+                    allow_face_overlap: bool = False,
+                    show_branch_length: bool = False,
+                    show_branch_support: bool = False,
                     tree_scale: float = None) -> TreeStyler:
         """
         Constructs a new :py:class:`genomics_data_index.api.viewer.TreeStyler` object used to style and visualize trees.
@@ -183,6 +231,7 @@ class TreeSamplesQuery(WrappedSamplesQuery, abc.ABC):
         :param highlight_style: A style used to define how the highlight() method should behave.
                                 Can either be one of the named highlight styles ['light', 'light_hn', 'pastel', 'medium', dark']
                                 or an instance of a :py:class:`genomics_data_index.api.viewer.TreeStyler.HighlightStyle`.
+        :param node_style: The default ete3.NodeStyle object for styling the nodes of the tree.
         :param legend_nsize: The legend node size.
         :param legend_fsize: The legend font size.
         :param annotate_color_present: The default color of samples which are present in the set for the annotate() method.
@@ -214,6 +263,10 @@ class TreeSamplesQuery(WrappedSamplesQuery, abc.ABC):
         :param show_legend_type_labels: Whether or not to show labels for legend types/categories (present or unknown).
         :param legend_type_label_present: Text to show above legend color for present items.
         :param legend_type_label_unknown: Text to show above legend color for unknown items.
+        :param rotation: The rotation of the tree in degrees (for circular mode).
+        :param allow_face_overlap: Allow overlap in node faces for circular images.
+        :param show_branch_length: Show branch lengths.
+        :param show_branch_support: Show branch supports.
         :param tree_scale: A scale factor for the tree.
         :return: A new :py:class:`genomics_data_index.api.viewer.TreeStyler` object used to style and visualize trees.
         """
@@ -221,6 +274,7 @@ class TreeSamplesQuery(WrappedSamplesQuery, abc.ABC):
                                  initial_style=initial_style,
                                  mode=mode,
                                  highlight_style=highlight_style,
+                                 node_style=node_style,
                                  legend_nsize=legend_nsize,
                                  legend_fsize=legend_fsize,
                                  annotate_color_present=annotate_color_present,
@@ -251,6 +305,10 @@ class TreeSamplesQuery(WrappedSamplesQuery, abc.ABC):
                                  show_legend_type_labels=show_legend_type_labels,
                                  legend_type_label_present=legend_type_label_present,
                                  legend_type_label_unknown=legend_type_label_unknown,
+                                 rotation=rotation,
+                                 allow_face_overlap=allow_face_overlap,
+                                 show_branch_length=show_branch_length,
+                                 show_branch_support=show_branch_support,
                                  tree_scale=tree_scale)
 
     @property