+++ genomics_data_index/api/viewer/TreeStyler.py
@@ -2,7 +2,7 @@ from __future__ import annotations
 
 import copy
 import logging
-from typing import List, Dict, Any, Union, Iterable
+from typing import List, Dict, Any, Union, Iterable, Tuple
 
 from ete3 import Tree, NodeStyle, TreeStyle, TextFace, RectFace
 
@@ -12,6 +12,9 @@ from genomics_data_index.api.viewer.samples_visuals.AnnotateTreeSamplesVisual im
 from genomics_data_index.api.viewer.samples_visuals.HighlightTreeSamplesVisual import HighlightTreeSamplesVisual
 
 logger = logging.getLogger(__name__)
+
+DEFAULT_NODE_STYLE = NodeStyle()
+DEFAULT_NODE_STYLE['size'] = 0
 
 
 class TreeStyler:
@@ -23,6 +26,7 @@ class TreeStyler:
 
     def __init__(self, tree: Tree, default_highlight_styles: HighlightStyle, annotate_column: int,
                  tree_style: TreeStyle,
+                 node_style: NodeStyle,
                  legend_columns: Dict[str, int],
                  samples_styles_list: List[TreeSamplesVisual],
                  legend_nsize: int = 20, legend_fsize: int = 11,
@@ -50,6 +54,7 @@ class TreeStyler:
                                 or an instance of a :py:class:`genomics_data_index.api.viewer.TreeStyler.HighlightStyle`.
         :param annotate_column: Which column in the annotate table this annotate visual belongs to (starting from column 0).
         :param tree_style: The ete3.TreeStyle object defining the style of the tree.
+        :param node_style: The default ete3.NodeStyle object for styling the nodes of the tree.
         :param legend_columns: A dictionary mapping legend column names to integers defining the column numbers.
         :param samples_styles_list: A list of :py:class:`genomics_data_index.api.viewer.TreeSamplesVisual` objects
                                    that define the different styles to apply to sets of samples (e.g., highlight or annotate).
@@ -76,6 +81,7 @@ class TreeStyler:
         self._tree = tree
         self._default_highlight_styles = default_highlight_styles
         self._tree_style = tree_style
+        self._node_style = node_style
         self._legend_columns = legend_columns
         self._legend_nsize = legend_nsize
         self._legend_fsize = legend_fsize
@@ -206,6 +212,7 @@ class TreeStyler:
 
         return TreeStyler(self._tree, default_highlight_styles=self._default_highlight_styles,
                           tree_style=self._tree_style,
+                          node_style=self._node_style,
                           samples_styles_list=samples_styles_list_new,
                           legend_fsize=self._legend_fsize, legend_nsize=self._legend_nsize,
                           annotate_column=self._annotate_column + 1,
@@ -270,6 +277,7 @@ class TreeStyler:
 
         return TreeStyler(self._tree, default_highlight_styles=new_default_styles,
                           tree_style=self._tree_style,
+                          node_style=self._node_style,
                           samples_styles_list=samples_styles_list_new,
                           legend_fsize=self._legend_fsize, legend_nsize=self._legend_nsize,
                           annotate_column=self._annotate_column,
@@ -296,7 +304,8 @@ class TreeStyler:
             samples_style.apply_visual(tree=tree, tree_style=tree_style)
 
     def render(self, file_name: str = '%%inline', w: int = None, h: int = None,
-               tree_style: TreeStyle = None, units: str = 'px', dpi: int = 90):
+               tree_style: TreeStyle = None, units: str = 'px', dpi: int = 90,
+               ladderize: bool = False):
         """
         Renders the tree with the styles defined by this TreeStyler object to an image.
         :param file_name: The image file name to save, use '%%inline' to draw inline in Jupyter notebooks (default).
@@ -305,10 +314,9 @@ class TreeStyler:
         :param tree_style: The ete3.TreeStyle object (overrides the object defined by this TreeStyler).
         :param units: The units of the width/height (default in pixels).
         :param dpi: The dots per inch.
+        :param ladderize: Runs ete3.Tree.ladderize() to sort branches of a given tree according to size.
         :return: An image/image data of the drawn tree styled according to this TreeStyler.
         """
-        if tree_style is None:
-            tree_style = self._tree_style
 
         # Set default width if no width or height specified
         # Do this here instead of as a default method value since at least one of
@@ -317,10 +325,51 @@ class TreeStyler:
         if w is None and h is None:
             w = 400
 
+        tree, tree_style = self.prerender(tree_style=tree_style, ladderize=ladderize)
+
+        return self.draw(tree=tree, tree_style=tree_style,
+                         file_name=file_name, w=w, h=h,
+                         units=units, dpi=dpi)
+
+    def prerender(self, tree_style: TreeStyle = None, ladderize: bool = False) -> Tuple[Tree, TreeStyle]:
+        """
+        Pre-renders the tree, returning an ete3.Tree and ete3.TreeStyle object which can be used to draw the tree.
+        :param tree_style: The ete3.TreeStyle object (overrides the object defined by this TreeStyler).
+        :param ladderize: Runs ete3.Tree.ladderize() to sort branches of a given tree according to size.
+        :return: A tuple of <ete3.Tree, ete3.TreeStyle> with highlights/annotations added.
+        """
+        if tree_style is None:
+            tree_style = self._tree_style
+
         tree = self._tree.copy('newick')
         tree_style = copy.deepcopy(tree_style)
+
+        # Sets default node style for all nodes in the tree
+        for n in tree.traverse():
+            n.set_style(self._node_style)
+
         self._apply_samples_styles(tree=tree, tree_style=tree_style)
 
+        if ladderize:
+            tree.ladderize()
+
+        return tree, tree_style
+
+    @classmethod
+    def draw(cls, tree: Tree, tree_style: TreeStyle,
+             file_name: str = '%%inline', w: int = None, h: int = None,
+             units: str = 'px', dpi: int = 90):
+        """
+        Draws the passed tree with the passed styles to an image.
+        :param tree: The tree to draw.
+        :param tree_style: The ete3.TreeStyle object (overrides the object defined by this TreeStyler).
+        :param file_name: The image file name to save, use '%%inline' to draw inline in Jupyter notebooks (default).
+        :param w: The width of the image.
+        :param h: The height of the image.
+        :param units: The units of the width/height (default in pixels).
+        :param dpi: The dots per inch.
+        :return: An image/image data of the drawn tree styled according to this TreeStyler.
+        """
         return tree.render(file_name=file_name, w=w, h=h, tree_style=tree_style,
                            units=units, dpi=dpi)
 
@@ -348,8 +397,9 @@ class TreeStyler:
     @classmethod
     def create(cls, tree: Tree,
                initial_style: TreeStyle = None,
-               mode='r',
+               mode: str = 'r',
                highlight_style: Union[str, HighlightStyle] = 'light',
+               node_style: NodeStyle = None,
                legend_nsize: int = 20, legend_fsize: int = 11,
                annotate_color_present: str = 'black',
                annotate_color_absent: str = 'white',
@@ -379,6 +429,10 @@ class TreeStyler:
                show_legend_type_labels: bool = True,
                legend_type_label_present: str = 'Pr.',
                legend_type_label_unknown: str = 'Un.',
+               rotation: float = 0,
+               allow_face_overlap: bool = False,
+               show_branch_length: bool = False,
+               show_branch_support: bool = False,
                tree_scale: float = None) -> TreeStyler:
         """
         Constructs a new TreeStyler object used to style and visualize trees.
@@ -389,6 +443,7 @@ class TreeStyler:
         :param highlight_style: A style used to define how the highlight() method should behave.
                                 Can either be one of the named highlight styles ['light', 'light_hn', 'pastel', 'medium', dark']
                                 or an instance of a :py:class:`genomics_data_index.api.viewer.TreeStyler.HighlightStyle`.
+        :param node_style: The default ete3.NodeStyle object for styling the nodes of the tree.
         :param legend_nsize: The legend node size.
         :param legend_fsize: The legend font size.
         :param annotate_color_present: The default color of samples which are present in the set for the annotate() method.
@@ -420,6 +475,10 @@ class TreeStyler:
         :param show_legend_type_labels: Whether or not to show labels for legend types/categories (present or unknown).
         :param legend_type_label_present: Text to show above legend color for present items.
         :param legend_type_label_unknown: Text to show above legend color for unknown items.
+        :param rotation: The rotation of the tree in degrees (for circular mode).
+        :param allow_face_overlap: Allow overlap in node faces for circular images.
+        :param show_branch_length: Show branch lengths.
+        :param show_branch_support: Show branch supports.
         :param tree_scale: A scale factor for the tree.
         :return: A new TreeStyler object used to style and visualize trees.
         """
@@ -428,7 +487,9 @@ class TreeStyler:
                                    'figure_margin': figure_margin, 'show_border': show_border,
                                    'title': title, 'legend_title': legend_title, 'title_fsize': title_fsize,
                                    'annotate_arc_span': annotate_arc_span, 'show_leaf_names': show_leaf_names,
-                                   'tree_scale': tree_scale}
+                                   'tree_scale': tree_scale, 'rotation': rotation,
+                                   'allow_face_overlap': allow_face_overlap, 'show_branch_length': show_branch_length,
+                                   'show_branch_support': show_branch_support}
             tree_style_elements_none = [tree_style_elements[k] is None for k in tree_style_elements]
 
             if any(tree_style_elements_none):
@@ -440,6 +501,10 @@ class TreeStyler:
             ts = TreeStyle()
             ts.arc_span = annotate_arc_span
             ts.show_leaf_name = show_leaf_names
+            ts.rotation = rotation
+            ts.allow_face_overlap = allow_face_overlap
+            ts.show_branch_length = show_branch_length
+            ts.show_branch_support = show_branch_support
 
             if mode is not None and mode not in cls.MODES:
                 raise Exception(f'Invalid value mode=[{mode}]. Must be one of {cls.MODES}')
@@ -471,10 +536,13 @@ class TreeStyler:
                 'text': 1
             }
 
+        if node_style is None:
+            node_style = DEFAULT_NODE_STYLE
+
         if highlight_style is None:
-            highlight_style = HighlightStyle.create('pastel')
+            highlight_style = HighlightStyle.create('pastel', base_node_style=node_style)
         elif isinstance(highlight_style, str):
-            highlight_style = HighlightStyle.create(highlight_style)
+            highlight_style = HighlightStyle.create(highlight_style, base_node_style=node_style)
 
         if legend_title is not None:
             margin_bottom = 10
@@ -518,6 +586,7 @@ class TreeStyler:
         return TreeStyler(tree=tree,
                           default_highlight_styles=highlight_style,
                           tree_style=ts,
+                          node_style=node_style,
                           legend_columns=legend_columns,
                           samples_styles_list=[],
                           legend_nsize=legend_nsize,
@@ -595,59 +664,74 @@ class HighlightStyle:
         return HighlightStyle(self._node_styles, index=new_index)
 
     @classmethod
-    def create(cls, kind: str) -> HighlightStyle:
+    def create(cls, kind: str = None, colors: List[str] = None, base_node_style: NodeStyle = None) -> HighlightStyle:
         """
         Creates a new pre-defined HighlightStyle.
         :param kind: The kind (name) of the pre-defined HighlightStyle to create.
+        :param colors: A list of colors to use for highlights (unused if kind is set).
+        :param base_node_style: The base node style to use.
         :return: A new HighlightStyle.
         """
-        unknown_fg_color = 'lightgray'
         unknown_bg_color = 'lightgray'
 
+        if base_node_style is None:
+            base_node_style = DEFAULT_NODE_STYLE
+
+        default_fg_color = base_node_style['fgcolor']
+        unknown_fg_color = default_fg_color
+
+        if kind is None and colors is not None:
+            return cls._create_highlights(base_node_style=base_node_style, fg_colors=[default_fg_color] * len(colors),
+                                          bg_colors=colors, unknown_bg_color=unknown_bg_color,
+                                          unknown_fg_color=unknown_fg_color)
+        elif kind is None:
+            kind = 'light'
+
         if kind == 'light':
-            fg_colors = ['#e5f5f9', '#fee8c8', '#e0ecf4', '#deebf7']
-            bg_colors = fg_colors
-            return cls._create_highlights(fg_colors=fg_colors, bg_colors=bg_colors,
+            fg_colors = [default_fg_color] * 4
+            bg_colors = ['#e5f5f9', '#fee8c8', '#e0ecf4', '#deebf7']
+            return cls._create_highlights(base_node_style=base_node_style, fg_colors=fg_colors, bg_colors=bg_colors,
                                           unknown_bg_color=unknown_bg_color, unknown_fg_color=unknown_fg_color)
         elif kind == 'light_hn':
             fg_colors = ['#41ae76', '#ef6548', '#8c6bb1', '#4292c6']
             bg_colors = ['#e5f5f9', '#fee8c8', '#e0ecf4', '#deebf7']
-            return cls._create_highlights(fg_colors=fg_colors, bg_colors=bg_colors, nsize=20,
+            return cls._create_highlights(base_node_style=base_node_style, fg_colors=fg_colors, bg_colors=bg_colors,
+                                          nsize=20,
                                           unknown_bg_color=unknown_bg_color, unknown_fg_color=unknown_fg_color)
         elif kind == 'pastel':
-            fg_colors = ['#fbb4ae', '#b3cde3', '#ccebc5', '#decbe4', '#fed9a6']
-            bg_colors = fg_colors
-            return cls._create_highlights(fg_colors=fg_colors, bg_colors=bg_colors,
+            fg_colors = [default_fg_color] * 5
+            bg_colors = ['#fbb4ae', '#b3cde3', '#ccebc5', '#decbe4', '#fed9a6']
+            return cls._create_highlights(base_node_style=base_node_style, fg_colors=fg_colors, bg_colors=bg_colors,
                                           unknown_bg_color=unknown_bg_color, unknown_fg_color=unknown_fg_color)
         elif kind == 'medium':
-            fg_colors = ['#66c2a5', '#fc8d62', '#8da0cb', '#e78ac3', '#a6d854']
-            bg_colors = fg_colors
-            return cls._create_highlights(fg_colors=fg_colors, bg_colors=bg_colors,
+            fg_colors = [default_fg_color] * 5
+            bg_colors = ['#66c2a5', '#fc8d62', '#8da0cb', '#e78ac3', '#a6d854']
+            return cls._create_highlights(base_node_style=base_node_style, fg_colors=fg_colors, bg_colors=bg_colors,
                                           unknown_bg_color=unknown_bg_color, unknown_fg_color=unknown_fg_color)
         elif kind == 'dark':
-            fg_colors = ['#1b9e77', '#d95f02', '#7570b3', '#e7298a', '#66a61e']
-            bg_colors = fg_colors
-            return cls._create_highlights(fg_colors=fg_colors, bg_colors=bg_colors,
+            fg_colors = [default_fg_color] * 5
+            bg_colors = ['#1b9e77', '#d95f02', '#7570b3', '#e7298a', '#66a61e']
+            return cls._create_highlights(base_node_style=base_node_style, fg_colors=fg_colors, bg_colors=bg_colors,
                                           unknown_bg_color=unknown_bg_color, unknown_fg_color=unknown_fg_color)
         else:
             raise Exception(f'kind=[{kind}] must be one of {cls.THEMES}')
 
     @classmethod
-    def _create_highlights(cls, fg_colors: List[str], bg_colors: List[str],
+    def _create_highlights(cls, base_node_style: NodeStyle, fg_colors: List[str], bg_colors: List[str],
                            unknown_fg_color: str, unknown_bg_color: str,
                            nsize: int = None) -> HighlightStyle:
         if len(fg_colors) != len(bg_colors):
             raise Exception(f'fg_colors={fg_colors} and bg_colors={bg_colors} do not have the same length')
 
         styles = []
-        for i in range(0, len(fg_colors)):
-            nstyle = NodeStyle()
+        for i in range(0, len(bg_colors)):
+            nstyle = copy.deepcopy(base_node_style)
             nstyle['fgcolor'] = fg_colors[i]
             nstyle['bgcolor'] = bg_colors[i]
             if nsize is not None:
                 nstyle['size'] = nsize
 
-            unknown_nstyle = NodeStyle()
+            unknown_nstyle = copy.deepcopy(base_node_style)
             unknown_nstyle['fgcolor'] = unknown_fg_color
             unknown_nstyle['bgcolor'] = unknown_bg_color
             if nsize is not None:
@@ -656,7 +740,7 @@ class HighlightStyle:
             styles.append({
                 'present_nstyle': nstyle,
                 'unknown_nstyle': unknown_nstyle,
-                'legend_color': fg_colors[i]
+                'legend_color': bg_colors[i]
             })
 
         return HighlightStyle(styles, index=0)