+++ CHANGELOG.md
@@ -1,6 +1,15 @@
 # 0.4.0
 
 * [analysis]: Switched all steps to use conda in Snakemake pipeline (0.3.1).
+* [api]: Added `rotation`, `allow_face_overlap`, `show_branch_length`, and `show_branch_support` to tree viewing API (0.4.0.dev1).
+* [api]: Fixed up `features_comparison` for joined dataframe to query so it uses proper subset of dataframe (0.4.0.dev1).
+* [api]: Implemented a `prune()` method for pruning a tree down to only the selected samples (0.4.0.dev1).
+* [api]: Fixed bug where `query.join_tree()` was pruning the original tree (instead of a copy) (0.4.0.dev1).
+* [cli]: Fixed bug where `--features-summary mlst` would not work in command-line interface (0.4.0.dev1).
+* [api]: Changed default NodeStyle for rendering trees such that nodes have size 0, which avoids inflating distances when many samples have distance 0 (0.4.0.dev2).
+* [api]: Adding ability to more easily set highlight colours and adjusted default node colours for highlights (0.4.0.dev2).
+* [api]: Added ability to pre-render a tree and included additional parameters for rendering (0.4.0.dev2).
+* [analysis]: Fixed issue where incorrect snpEff annotation was being loaded for ORF1ab in SARS-CoV-2 (0.4.0.dev3).
 
 # 0.3.0